backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.1
beanie==1.25.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
cryptography==41.0.7
dnspython==2.6.1
ecdsa==0.18.0
fastapi==0.104.1
h11==0.14.0
html5lib==1.1
httptools==0.6.1
humanfriendly==10.0
idna==3.6
isodate==0.6.1
kink==0.7.0
lazy-model==0.2.0
motor==3.3.2
numpy==1.26.4
owlrl==6.0.2
packaging==24.0
pandas==2.2.1
prettytable==3.10.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.2
pyparsing==3.1.1
pyshacl==0.25.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
rdflib==7.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
